Av1-Backend/requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & API
djangorestframework-simplejwt==5.3.0

# Filtering
django-filter==23.4

# Build tools (to fix pkg_resources deprecation warning)
setuptools>=68.0.0

# Image processing
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# Excel Generation
openpyxl>=3.1.0

# Database connectors (elige uno según tu BD)
mysqlclient==2.2.0  # Para MySQL
# psycopg2-binary==2.9.9  # Para PostgreSQL